# server/src/openapi/openapi.yaml
...
openapi: 3.0.3
info:
  #  title: Mentor Success API
  description: API for the Mentor Success application
  contact:
    email: whitney@aiontechnology.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.1.0
servers:
  - url: "https://lj6z99n04d.execute-api.us-west-2.amazonaws.com/mentorsuccess-test-service-deployment"
...
paths:
...
  /api/v1/schools/{schoolId}:
    parameters:
      - in: header
        name: Authorization
        schema:
          type: string
      - in: path
        name: schoolId
        schema:
          type: string
          format: uuid
        required: true
        description: The id of the desired school

    # OPTIONS for /api/v1/schools/{schoolId}
    options:
      tags:
        - school
      responses:
        200:
          description: Success
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: "string"
            Access-Control-Allow-Methods:
              schema:
                type: "string"
            Access-Control-Allow-Headers:
              schema:
                type: "string"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-amazon-apigateway-auth:
        type: None
      x-amazon-apigateway-integration:
        responses:
          default:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Methods: "'DELETE,GET,HEAD,OPTIONS,PATCH,POST,PUT'"
              method.response.header.Access-Control-Allow-Headers: "'Content-Type,Authorization,X-Amz-Date,X-Api-Key,X-Amz-Security-Token'"
              method.response.header.Access-Control-Allow-Origin: "'*'"
        requestTemplates:
          application/json: "{\"statusCode\": 200}"
        passthroughBehavior: "when_no_match"
        type: "mock"

    # GET for /api/v1/schools/{schoolId}
    get:
      summary: Get a school
      description: Get a particular school by its id
      tags:
        - school
      operationId: getSchool
      responses:
        200:
          description: The school was found
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/NotAuthenticated'
        403:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/Error'
      x-amazon-apigateway-auth:
        type: None
      x-amazon-apigateway-integration:
        uri: "${lb_url}/api/v1/schools/{schoolId}"
        responses:
          200:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          400:
            statusCode: "400"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          401:
            statusCode: "401"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          403:
            statusCode: "403"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          404:
            statusCode: "404"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          default:
            statusCode: "500"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
        requestParameters:
          integration.request.header.Authorization: "method.request.header.Authorization"
          integration.request.header.X-Forwarded-Proto: "'https'"
          integration.request.path.schoolId: "method.request.path.schoolId"
          integration.request.header.X-Forwarded-Host: "stageVariables.api_url"
        passthroughBehavior: "when_no_match"
        connectionType: "VPC_LINK"
        connectionId: "$${stageVariables.vpc_link_id}"
        httpMethod: "GET"
        type: "http"

    # DELETE for /api/v1/schools/{schoolId}
    delete:
      summary: Delete a school
      description: Delete the school with the given id
      tags:
        - school
      operationId: removeSchool
      responses:
        204:
          description: The school was deleted
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/NotAuthenticated'
        403:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/Error'
      x-amazon-apigateway-auth:
        type: None
      x-amazon-apigateway-integration:
        uri: "${lb_url}/api/v1/schools/{schoolId}"
        responses:
          204:
            statusCode: "204"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          400:
            statusCode: "400"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          401:
            statusCode: "401"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          403:
            statusCode: "403"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          404:
            statusCode: "404"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          default:
            statusCode: "500"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
        requestParameters:
          integration.request.header.Authorization: "method.request.header.Authorization"
          integration.request.header.X-Forwarded-Proto: "'https'"
          integration.request.path.schoolId: "method.request.path.schoolId"
          integration.request.header.X-Forwarded-Host: "stageVariables.api_url"
        passthroughBehavior: "when_no_match"
        connectionType: "VPC_LINK"
        connectionId: "$${stageVariables.vpc_link_id}"
        httpMethod: "DELETE"
        type: "http"

    # PUT for /api/v1/schools/{schoolId}
    put:
      summary: Update a school
      description: Update the school with the given id
      tags:
        - school
      operationId: updateSchool
      parameters:
        - in: path
          name: schoolId
          schema:
            type: string
            format: uuid
          required: true
          description: The id of the desired school
      requestBody:
        description: The updated school
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchoolModel'
      responses:
        200:
          description: The school was updated
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchoolModel'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/NotAuthenticated'
        403:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/Error'
      x-amazon-apigateway-auth:
        type: None
      x-amazon-apigateway-integration:
        uri: "${lb_url}/api/v1/schools/{schoolId}"
        responses:
          200:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          400:
            statusCode: "400"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          401:
            statusCode: "401"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          403:
            statusCode: "403"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          404:
            statusCode: "404"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          default:
            statusCode: "500"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
        requestParameters:
          integration.request.header.Authorization: "method.request.header.Authorization"
          integration.request.header.X-Forwarded-Proto: "'https'"
          integration.request.path.schoolId: "method.request.path.schoolId"
          integration.request.header.X-Forwarded-Host: "stageVariables.api_url"
        passthroughBehavior: "when_no_match"
        connectionType: "VPC_LINK"
        connectionId: "$${stageVariables.vpc_link_id}"
        httpMethod: "PUT"
        type: "http"

  # /api/v1/schools/{schoolId}/books -----------------------------------------------
...
  /api/v1/schools/{schoolId}/schoolsessions:
    parameters:
      - in: header
        name: Authorization
        schema:
          type: string
      - in: path
        name: schoolId
        schema:
          type: string
          format: uuid
        required: true
        description: The id of the desired school

    # OPTIONS for /api/v1/schools/{schoolId}/schoolsessions
    options:
      tags:
        - school
      responses:
        200:
          description: Success
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: "string"
            Access-Control-Allow-Methods:
              schema:
                type: "string"
            Access-Control-Allow-Headers:
              schema:
                type: "string"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-amazon-apigateway-auth:
        type: None
      x-amazon-apigateway-integration:
        responses:
          default:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Methods: "'DELETE,GET,HEAD,OPTIONS,PATCH,POST,PUT'"
              method.response.header.Access-Control-Allow-Headers: "'Content-Type,Authorization,X-Amz-Date,X-Api-Key,X-Amz-Security-Token'"
              method.response.header.Access-Control-Allow-Origin: "'*'"
        requestTemplates:
          application/json: "{\"statusCode\": 200}"
        passthroughBehavior: "when_no_match"
        type: "mock"

    # GET for /api/v1/schools/{schoolId}/schoolsessions
    get:
      summary: Get a school session
      description: Get the sessions for a particular school by its id
      tags:
        - school
      operationId: getSchoolSessions
      responses:
        200:
          description: The school sessions were found
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/NotAuthenticated'
        403:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/Error'
      x-amazon-apigateway-auth:
        type: None
      x-amazon-apigateway-integration:
        uri: "${lb_url}/api/v1/schools/{schoolId}/schoolsessions"
        responses:
          200:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          400:
            statusCode: "400"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          401:
            statusCode: "401"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          403:
            statusCode: "403"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          404:
            statusCode: "404"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          default:
            statusCode: "500"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
        requestParameters:
          integration.request.header.Authorization: "method.request.header.Authorization"
          integration.request.header.X-Forwarded-Proto: "'https'"
          integration.request.path.schoolId: "method.request.path.schoolId"
          integration.request.header.X-Forwarded-Host: "stageVariables.api_url"
        passthroughBehavior: "when_no_match"
        connectionType: "VPC_LINK"
        connectionId: "$${stageVariables.vpc_link_id}"
        httpMethod: "GET"
        type: "http"

    # POST for /api/v1/schools/{schoolId}/schoolsessions
    post:
      summary: Create a new school session
      description: Create a new school session within the system
      tags:
        - school
      operationId: createSchoolSession
      responses:
        201:
          description: A new school session was created
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchoolSessionResultModel'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/NotAuthenticated'
        403:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/Error'
      x-amazon-apigateway-auth:
        type: None
      x-amazon-apigateway-integration:
        uri: "${lb_url}/api/v1/schools/{schoolId}/schoolsessions"
        responses:
          201:
            statusCode: "201"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          401:
            statusCode: "401"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          403:
            statusCode: "403"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          404:
            statusCode: "404"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          default:
            statusCode: "500"
            description: Unable to process request
        requestParameters:
          integration.request.header.Authorization: "method.request.header.Authorization"
          integration.request.header.X-Forwarded-Proto: "'https'"
          integration.request.path.schoolId: "method.request.path.schoolId"
          integration.request.header.X-Forwarded-Host: "stageVariables.api_url"
        passthroughBehavior: "when_no_match"
        connectionType: "VPC_LINK"
        connectionId: "$${stageVariables.vpc_link_id}"
        httpMethod: "POST"
        type: "http"

  # /api/v1/schools/{schoolId}/schoolsessions/{schoolSessionId} --------------------
  /api/v1/schools/{schoolId}/schoolsessions/{schoolSessionId}:
    parameters:
      - in: header
        name: Authorization
        schema:
          type: string
      - in: path
        name: schoolId
        schema:
          type: string
          format: uuid
        required: true
        description: The id of the school that owns the session
      - in: path
        name: schoolSessionId
        schema:
          type: string
          format: uuid
        required: true
        description: The id of the desired school session

    # OPTIONS for /api/v1/schools/{schoolId}/schoolsessions/{schoolSessionId}
    options:
      tags:
        - school
      responses:
        200:
          description: Success
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: "string"
            Access-Control-Allow-Methods:
              schema:
                type: "string"
            Access-Control-Allow-Headers:
              schema:
                type: "string"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-amazon-apigateway-auth:
        type: None
      x-amazon-apigateway-integration:
        responses:
          default:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Methods: "'DELETE,GET,HEAD,OPTIONS,PATCH,POST,PUT'"
              method.response.header.Access-Control-Allow-Headers: "'Content-Type,Authorization,X-Amz-Date,X-Api-Key,X-Amz-Security-Token'"
              method.response.header.Access-Control-Allow-Origin: "'*'"
        requestTemplates:
          application/json: "{\"statusCode\": 200}"
        passthroughBehavior: "when_no_match"
        type: "mock"

    # GET for /api/v1/schools/{schoolId}/schoolsessions/{schoolSessionId}
    get:
      summary: Get a school session
      description: Get a particular school session by its id
      tags:
        - school
      operationId: getSchoolSession
      responses:
        200:
          description: The school sessions was found
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/NotAuthenticated'
        403:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/Error'
      x-amazon-apigateway-auth:
        type: None
      x-amazon-apigateway-integration:
        uri: "${lb_url}/api/v1/schools/{schoolId}/schoolsessions/{schoolSessionId}"
        responses:
          200:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          400:
            statusCode: "400"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          401:
            statusCode: "401"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          403:
            statusCode: "403"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          404:
            statusCode: "404"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          default:
            statusCode: "500"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
        requestParameters:
          integration.request.header.Authorization: "method.request.header.Authorization"
          integration.request.header.X-Forwarded-Proto: "'https'"
          integration.request.path.schoolId: "method.request.path.schoolId"
          integration.request.path.schoolSessionId: "method.request.path.schoolSessionId"
          integration.request.header.X-Forwarded-Host: "stageVariables.api_url"
        passthroughBehavior: "when_no_match"
        connectionType: "VPC_LINK"
        connectionId: "$${stageVariables.vpc_link_id}"
        httpMethod: "GET"
        type: "http"

    # DELETE for /api/v1/schools/{schoolId}/schoolsessions/{schoolSessionId}
    delete:
      summary: Disable a school session
      description: Disable a school session record
      tags:
        - school
      operationId: deactivateSchoolSession
      responses:
        204:
          description: The school session record was deleted
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/NotAuthenticated'
        403:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/Error'
      x-amazon-apigateway-auth:
        type: None
      x-amazon-apigateway-integration:
        uri: "${lb_url}/api/v1/schools/{schoolId}/schoolsessions/{schoolSessionId}"
        responses:
          204:
            statusCode: "204"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          400:
            statusCode: "400"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          401:
            statusCode: "401"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          403:
            statusCode: "403"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          404:
            statusCode: "404"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          default:
            statusCode: "500"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
        requestParameters:
          integration.request.header.Authorization: "method.request.header.Authorization"
          integration.request.header.X-Forwarded-Proto: "'https'"
          integration.request.path.schoolId: "method.request.path.schoolId"
          integration.request.path.schoolSessionId: "method.request.path.schoolSessionId"
          integration.request.header.X-Forwarded-Host: "stageVariables.api_url"
        passthroughBehavior: "when_no_match"
        connectionType: "VPC_LINK"
        connectionId: "$${stageVariables.vpc_link_id}"
        httpMethod: "DELETE"
        type: "http"

  # /api/v1/schools/{schoolId}/students --------------------------------------------
...
components:
  responses:
    BadRequest:
      description: There was an error while creating the person
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BadRequestModel'
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
    Error:
      description: Encountered a server error
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
    NotAuthenticated:
      description: Access token is missing or invalid
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
    NotAuthorized:
      description: User does not have permissions for operation
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
    NotFound:
      description: Requested resource could not be found
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
...
  schemas:
...
    AddressModel:
      type: object
      properties:
        street1:
          type: string
          maxLength: 50
          example: 10818 NE 117th Ave
        street2:
          type: string
          maxLength: 50
        city:
          type: string
          maxLength: 50
          example: Vancouver
        state:
          type: string
          maxLength: 2
          example: WA
        zip:
          type: string
          maxLength: 9
          example: 98662

    # BadRequestModel --------------------------------------------------------------
    BadRequestModel:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: string
        error:
          type: object
        message:
          type: string
        path:
          type: string

    # BehaviorModelCollection ------------------------------------------------------
...
    Empty:
      type: object

    # GameModel --------------------------------------------------------------------
...
    SchoolModel:
      required:
        - name
      properties:
        id:
          type: string
          format: uuid
          example: 734aa3e8-52e4-488b-8922-09d8eb803ad8
        name:
          type: string
          maxLength: 50
          example: Cornerstone Christian Acadamy
        address:
          type: object
          allOf:
            - $ref: '#/components/schemas/AddressModel'
        phone:
          type: string
          maxLength: 10
          example: 3602569715
        district:
          type: string
          maxLength: 50
          example: Vancouver
        isPrivate:
          type: boolean
          example: true

    # SchoolResultModel ------------------------------------------------------------
...
    SchoolSessionModel:
      required:
        - name
      properties:
        startDate:
          type: string
          format: date-time
          example: 01/01/2022
        endDate:
          type: string
          format: date-time
          example: 01/01/2022
        label:
          type: string
          maxLength: 30
          example: 2022-23

    # SchoolSessionResultModel -----------------------------------------------------
    SchoolSessionResultModel:
      type: object
      allOf:
        - $ref: '#/components/schemas/SchoolSessionModel'
      properties:
        _links:
          type: object
          allOf:
            - $ref: '#/components/schemas/SelfLink'

    # SchoolSessionResultModelCollection -------------------------------------------
...
    SelfLink:
      type: object
      properties:
        self:
          allOf:
            - $ref: '#/components/schemas/UrlArray'

    # StudentModel -----------------------------------------------------------------
...
    UrlArray:
      type: array
      items:
        type: object
        properties:
          href:
            type: string
            format: url